Ref: Administration des modules totalement récrite Ref: Suppression des anciennes définitions de modules redondantes Bug: Correction de la dépendance post-installation à Mediusers Bug: La mise à jour du système de permissions ne fonctionnait plus suite au refactoring reflexif des ref_module
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@4356 a21d6be2-e512-0410-a060-91f5a2946631

+++ classes/setup.class.php
@@ -161,15 +161,22 @@ class CSetup {
   /**
    * Removes a module
    * Warning, it actually breaks module dependency
+   * @return bool Job done
    */
   function remove() {
     global $AppUI;
+    
+    if ($this->mod_type == "core") {
+      $AppUI->setMsg("Impossible de supprimer le module '%s'", UI_MSG_ERROR, $this->mod_name);
+      return false;
+    }
+    
     $success = true;
     foreach ($this->tables as $table) {
       $query = "DROP TABLE `$table`";
       if (!$this->ds->exec($query)) {
         $success = false;
-        $AppUI->setMsg("Failed to remove table '$table'", UI_MSG_ERROR, true);
+        $AppUI->setMsg("Failed to remove table '%s'", UI_MSG_ERROR, $table);
       } 
     }
